rename ScanType to AssessmentType and introduce IssueCalculator for issue handling
