+++ src/main/java/io/jenkins/plugins/security/scan/extension/pipeline/SecurityScanStep.java
@@ -1412,7 +1412,7 @@ public class SecurityScanStep extends Step
                             scanparametersMap
                                     .get(ApplicationConstants.PRODUCT_KEY)
                                     .toString(),
-                            Utility.getProductUrl(scanparametersMap),
+                            ParameterMappingService.getProductUrl(scanparametersMap),
                             workspace.child(ApplicationConstants.SCAN_INFO_OUT_FILE_NAME),
                             isPullRequestEvent));
                 }
