+++ src/main/java/io/jenkins/plugins/security/scan/global/enums/AssessmentType.java
@@ -1,6 +1,6 @@
 package io.jenkins.plugins.security.scan.global.enums;
 
-public enum ScanType {
+public enum AssessmentType {
     SAST,
     SCA;
 }
