+++ src/main/java/io/jenkins/plugins/security/scan/global/enums/TestScaType.java
@@ -1,0 +1,16 @@
+package io.jenkins.plugins.security.scan.global.enums;
+
+public enum TestScaType {
+    SCA_PACKAGE("scaPackage"),
+    SCA_SIGNATURE("scaSignature");
+
+    private final String value;
+
+    TestScaType(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
